requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
pymongo>=3.12.0
python-dotenv>=0.19.0
torch>=2.0.1
numpy>=1.21.2
pandas>=1.3.3
scikit-learn>=0.24.2
pytest>=6.2.5
gunicorn>=20.1.0
python-jose>=3.3.0
python-multipart>=0.0.5
nbformat
notebook